Address query cache helper review
