+++ tests/unit/QueryCacheTest.php
@@ -14,6 +14,8 @@ use Utopia\Database\Query;
 
 class QueryCacheTest extends TestCase
 {
+    use QueryCacheTestHelpers;
+
     /**
      * @param array<string, array{encode: callable, decode: callable}> $filters
      */
@@ -27,35 +29,6 @@ class QueryCacheTest extends TestCase
         $database->create();
 
         return $database;
-    }
-
-    private function purgeCachedQueries(Database $database, string $collection, ?string $namespace = null): bool
-    {
-        $method = new \ReflectionMethod(Database::class, 'purgeCachedQueries');
-
-        return $method->invoke($database, $collection, $namespace);
-    }
-
-    private function getQueryCacheKey(Database $database, string $collectionId, ?string $namespace = null): string
-    {
-        $method = new \ReflectionMethod(Database::class, 'getQueryCacheKey');
-
-        return $method->invoke($database, $collectionId, $namespace);
-    }
-
-    /**
-     * @param array<Query> $queries
-     */
-    private function getQueryCacheField(
-        Database $database,
-        ?Document $collection = null,
-        array $queries = [],
-        string $field = 'documents',
-        string $forPermission = Database::PERMISSION_READ,
-    ): ?string {
-        $method = new \ReflectionMethod(Database::class, 'getQueryCacheField');
-
-        return $method->invoke($database, $collection, $queries, $field, $forPermission);
     }
 
     /**
